chore(workspace): fix clippy warnings
All mechanical, behavior-free rewrites: derive Default instead of
manual impls, drop redundant f64::from on f64 values, let-chains for
nested ifs, iterator loops instead of index loops, range contains,
is_multiple_of, to_vec, struct-literal init in a test helper, and move
an item above its test module. update_sea_level gets an explicit
too_many_arguments allow, matching the existing allows in elevation.rs.

cargo clippy --workspace --all-targets -- -D warnings is now clean.

+++ crates/genesis_core/src/grid/mod.rs
@@ -152,9 +152,7 @@ impl HexGrid {
         let mut best = HexId(0);
         let mut best_dot = f64::NEG_INFINITY;
         for (index, center) in self.centers.iter().enumerate() {
-            let d = direction[0] * f64::from(center.x)
-                + direction[1] * f64::from(center.y)
-                + direction[2] * f64::from(center.z);
+            let d = direction[0] * center.x + direction[1] * center.y + direction[2] * center.z;
             let hex = HexId(index as u32);
             if d > best_dot {
                 best_dot = d;
@@ -216,9 +214,7 @@ impl HexGrid {
     /// Dot product of `hex`'s center direction with `direction`.
     fn dot_with_direction(&self, hex: HexId, direction: [f64; 3]) -> f64 {
         let center = self.centers[hex.0 as usize];
-        direction[0] * f64::from(center.x)
-            + direction[1] * f64::from(center.y)
-            + direction[2] * f64::from(center.z)
+        direction[0] * center.x + direction[1] * center.y + direction[2] * center.z
     }
 
     /// Iterate all valid [`HexId`] values in this grid (`0..cell_count`).
